fix debug ioctl call
When calling the ioctl to set a debug file, the ioctl expects to receive a datatype
that is specified in the ioctl definition: char[PATHLEN_MAX]

However when passing the pointer from getopts, this is only true if the passed path
has the maximal allowed length. Since usually this path is shorter than the max,
the kernel is trying to read over the end of the argument, and returns "EFAULT/Bad address".

To solve this, copy the argument to a buffer that has the exact size of what the ioctl
expects, and pass this buffer to the ioctl.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>